+++ batcher/src/builtin_procedures.py
@@ -325,6 +325,7 @@ def scale(
       interpolation,
       local_origin,
       aspect_ratio,
+      padding_color,
       set_image_resolution,
       image_resolution,
 ):
@@ -347,17 +348,15 @@ def scale(
     orig_height_pixels = 1
 
   if aspect_ratio in [AspectRatios.KEEP_ADJUST_WIDTH, AspectRatios.KEEP_ADJUST_HEIGHT]:
-    processed_width_pixels, processed_height_pixels = _get_keep_aspect_ratio_values(
+    processed_width_pixels, processed_height_pixels = _get_scale_keep_aspect_ratio_values(
       aspect_ratio,
       orig_width_pixels,
       orig_height_pixels,
       new_width_pixels,
       new_height_pixels)
-  elif aspect_ratio == AspectRatios.FIT:
-    processed_width_pixels, processed_height_pixels = _get_scale_to_fit_values(
+  elif aspect_ratio in [AspectRatios.FIT, AspectRatios.FIT_WITH_PADDING]:
+    processed_width_pixels, processed_height_pixels = _get_scale_fit_values(
       orig_width_pixels, orig_height_pixels, new_width_pixels, new_height_pixels)
-  elif aspect_ratio == AspectRatios.FIT_WITH_PADDING:
-    raise NotImplementedError
   else:
     processed_width_pixels = new_width_pixels
     processed_height_pixels = new_height_pixels
@@ -365,10 +364,25 @@ def scale(
   Gimp.context_push()
   Gimp.context_set_interpolation(interpolation)
 
+  if processed_width_pixels == 0:
+    processed_width_pixels = 1
+
+  if processed_height_pixels == 0:
+    processed_height_pixels = 1
+
   if isinstance(object_to_scale, Gimp.Image):
     object_to_scale.scale(processed_width_pixels, processed_height_pixels)
   else:
     object_to_scale.scale(processed_width_pixels, processed_height_pixels, local_origin)
+
+  if aspect_ratio == AspectRatios.FIT_WITH_PADDING:
+    _fill_with_padding(
+      batcher,
+      object_to_scale,
+      new_width_pixels,
+      new_height_pixels,
+      padding_color,
+    )
 
   Gimp.context_pop()
 
@@ -408,12 +422,13 @@ def _unit_to_pixels(batcher, dimension, dimension_name):
   return int_pixels
 
 
-def _get_keep_aspect_ratio_values(
+def _get_scale_keep_aspect_ratio_values(
       aspect_ratio,
       orig_width_pixels,
       orig_height_pixels,
       new_width_pixels,
-      new_height_pixels):
+      new_height_pixels,
+):
   if aspect_ratio == AspectRatios.KEEP_ADJUST_WIDTH:
     processed_new_width_pixels = new_width_pixels
     processed_new_height_pixels = round(
@@ -428,7 +443,7 @@ def _get_keep_aspect_ratio_values(
   return processed_new_width_pixels, processed_new_height_pixels
 
 
-def _get_scale_to_fit_values(
+def _get_scale_fit_values(
       orig_width_pixels, orig_height_pixels, new_width_pixels, new_height_pixels):
   processed_new_width_pixels = new_width_pixels
   processed_new_height_pixels = round(
@@ -440,6 +455,114 @@ def _get_scale_to_fit_values(
       orig_width_pixels * (new_height_pixels / orig_height_pixels))
 
   return processed_new_width_pixels, processed_new_height_pixels
+
+
+def _fill_with_padding(
+      batcher,
+      gimp_object,
+      new_width_pixels,
+      new_height_pixels,
+      padding_color,
+):
+  if isinstance(gimp_object, Gimp.Image):
+    drawable_with_padding = batcher.current_layer
+    image_of_drawable_with_padding = gimp_object
+  else:
+    drawable_with_padding = gimp_object
+    image_of_drawable_with_padding = gimp_object.get_image()
+
+  object_width = gimp_object.get_width()
+  object_height = gimp_object.get_height()
+
+  if new_width_pixels > object_width:
+    offset_x = (new_width_pixels - object_width) // 2
+    offset_y = 0
+    layer_to_fill_start_width = offset_x
+    layer_to_fill_start_height = new_height_pixels
+    layer_to_fill_end_width = offset_x + (new_width_pixels - object_width) % 2
+    layer_to_fill_end_height = new_height_pixels
+  else:
+    offset_x = 0
+    offset_y = (new_height_pixels - object_height) // 2
+    layer_to_fill_start_width = new_width_pixels
+    layer_to_fill_start_height = offset_y
+    layer_to_fill_end_width = new_width_pixels
+    layer_to_fill_end_height = offset_y + (new_height_pixels - object_height) % 2
+
+  if isinstance(gimp_object, Gimp.Image):
+    if new_width_pixels > object_width:
+      layer_to_fill_start_offset_x = 0
+      layer_to_fill_start_offset_y = 0
+      layer_to_fill_end_offset_x = offset_x + object_width
+      layer_to_fill_end_offset_y = offset_y
+    else:
+      layer_to_fill_start_offset_x = 0
+      layer_to_fill_start_offset_y = 0
+      layer_to_fill_end_offset_x = offset_x
+      layer_to_fill_end_offset_y = offset_y + object_height
+
+    gimp_object.resize(new_width_pixels, new_height_pixels, offset_x, offset_y)
+  else:
+    drawable_with_padding_offsets = gimp_object.get_offsets()
+
+    if new_width_pixels > object_width:
+      layer_to_fill_start_offset_x = drawable_with_padding_offsets.offset_x
+      layer_to_fill_start_offset_y = drawable_with_padding_offsets.offset_y
+      layer_to_fill_end_offset_x = drawable_with_padding_offsets.offset_x + object_width + offset_x
+      layer_to_fill_end_offset_y = drawable_with_padding_offsets.offset_y
+    else:
+      layer_to_fill_start_offset_x = drawable_with_padding_offsets.offset_x
+      layer_to_fill_start_offset_y = drawable_with_padding_offsets.offset_y
+      layer_to_fill_end_offset_x = drawable_with_padding_offsets.offset_x
+      layer_to_fill_end_offset_y = drawable_with_padding_offsets.offset_y + object_height + offset_y
+
+    gimp_object.transform_translate(offset_x, offset_y)
+
+  Gimp.context_set_foreground(pg.setting.ColorSetting.get_value_as_color(padding_color))
+  Gimp.context_set_opacity(
+    pg.setting.ColorSetting.get_value_as_color(padding_color).get_rgba().alpha * 100)
+
+  if layer_to_fill_start_width != 0 and layer_to_fill_start_height != 0:
+    layer_to_fill_start = Gimp.Layer.new(
+      image_of_drawable_with_padding,
+      drawable_with_padding.get_name(),
+      layer_to_fill_start_width,
+      layer_to_fill_start_height,
+      Gimp.ImageType.RGBA_IMAGE,
+      100.0,
+      Gimp.LayerMode.NORMAL,
+    )
+    layer_to_fill_start.set_offsets(layer_to_fill_start_offset_x, layer_to_fill_start_offset_y)
+    image_of_drawable_with_padding.insert_layer(
+      layer_to_fill_start,
+      drawable_with_padding.get_parent(),
+      image_of_drawable_with_padding.get_item_position(drawable_with_padding) + 1,
+    )
+    layer_to_fill_start.edit_fill(Gimp.FillType.FOREGROUND)
+    merged_drawable_with_padding = image_of_drawable_with_padding.merge_down(
+      drawable_with_padding, Gimp.MergeType.EXPAND_AS_NECESSARY)
+  else:
+    merged_drawable_with_padding = drawable_with_padding
+
+  if layer_to_fill_end_width != 0 and layer_to_fill_end_height != 0:
+    layer_to_fill_end = Gimp.Layer.new(
+      image_of_drawable_with_padding,
+      merged_drawable_with_padding.get_name(),
+      layer_to_fill_end_width,
+      layer_to_fill_end_height,
+      Gimp.ImageType.RGBA_IMAGE,
+      100.0,
+      Gimp.LayerMode.NORMAL,
+    )
+    layer_to_fill_end.set_offsets(layer_to_fill_end_offset_x, layer_to_fill_end_offset_y)
+    image_of_drawable_with_padding.insert_layer(
+      layer_to_fill_end,
+      merged_drawable_with_padding.get_parent(),
+      image_of_drawable_with_padding.get_item_position(merged_drawable_with_padding) + 1,
+    )
+    layer_to_fill_end.edit_fill(Gimp.FillType.FOREGROUND)
+    image_of_drawable_with_padding.merge_down(
+      merged_drawable_with_padding, Gimp.MergeType.EXPAND_AS_NECESSARY)
 
 
 class AspectRatios:
@@ -739,6 +862,12 @@ _SCALE_PROCEDURE_DICT_FOR_IMAGES = {
         (AspectRatios.FIT_WITH_PADDING, _('Fit with padding')),
       ],
       'display_name': _('Aspect ratio'),
+    },
+    {
+      'type': 'color',
+      'name': 'padding_color',
+      'default_value': [0.0, 0.0, 0.0, 0.0],
+      'display_name': _('Padding color'),
     },
     {
       'type': 'bool',